Fixed bugs in the Menu class by defining a class for validation
GUI should use the Check class to confirm User information before Login
