+++ AppDevProject/src/reservation/util/Test.java
@@ -5,22 +5,20 @@ public class Test {
 	public static void main(String[] args) {
 		
 		//test Login
-		
 		User user = null;
-		if (Menu.isValidUser("yomar", "123")) {
-			user = Menu.login("yomar", "123");
+		if (Check.isValidUser("yomar", "123")) {
+			user = Menu.login("ola", "ola123");
 		}
 		
-		if (user instanceof Customer) {
+		if (Check.isValidUser("yomar", "123") && (user instanceof Customer)) {
 			System.out.println(user.getFirstName());
 		}
 		
 		
-
 		
 		
 		// test registration
-		String[] newUser = {"tet","test1","Scot","Will","10102 Scot Rd","Atlanta","Georgia","30034",
+		String[] newUser = {"test","test1","Scot","Will","10102 Scot Rd","Atlanta","Georgia","30034",
 				"scot@yahoo.com","234567","Where were you born? ","Nigeria"};
 		
 		String response = Menu.register(newUser);
@@ -30,7 +28,7 @@ public class Test {
 		
 		
 		// test forgot password
-		String[] query = Menu.forgotPassword("test");
+		String[] query = Menu.forgotPassword("te");
 		System.out.println(query[0]);
